Fix: move to public folder
