+++ src/main/routes/features/public/eligibility/notEligibleController.ts
@@ -1,9 +1,9 @@
 import * as express from 'express';
-import {NOT_ELIGIBLE_FOR_THIS_SERVICE_URL} from '../../urls';
-import {convertToNotEligibleReason} from '../../../common/utils/notEligibleReasonConvertor';
+import {NOT_ELIGIBLE_FOR_THIS_SERVICE_URL} from '../../../urls';
+import {convertToNotEligibleReason} from '../../../../common/utils/notEligibleReasonConvertor';
 
 const notEligibleController = express.Router();
-const notEligibleControllerViewPath = 'features/eligibility/not-eligible';
+const notEligibleControllerViewPath = 'features/public/eligibility/not-eligible';
 
 notEligibleController
   .get(NOT_ELIGIBLE_FOR_THIS_SERVICE_URL, (req: express.Request, res: express.Response, next: express.NextFunction) => {
