+++ src/main/views/features/public/eligibility/not-eligible-includes/claim-value-over-25000.njk
@@ -1,4 +1,4 @@
-{% from "../../../macro/external-link.njk" import externalLink %}
+{% from "../../../../macro/external-link.njk" import externalLink %}
 {% import "external-urls.njk" as externalURLs %}
 
 <p class="govuk-body">{{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.FOR_CLAIMS_LESS_THAN_25000') }}</p>
