+++ src/main/views/features/public/eligibility/not-eligible-includes/defendant-address.njk
@@ -1,12 +1,12 @@
-{% from "../../../macro/external-link.njk" import externalLink %}
+{% from "../../../../macro/external-link.njk" import externalLink %}
 {% import "external-urls.njk" as externalURLs %}
 
 <p class="govuk-body">{{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.CAN_ONLY_USE_ADDRESS_ENGLAND_OR_WALES') }}</p>
 <p class="govuk-body">
   {{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.USING_PAPER_FORM') }} 
-  {{ externalLink(t, 'PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.DOWNLOAD_PAPER_FORM_N1', externalURLs.n1FormCprPart7Url, true) }}
+  {{ externalLink(t, 'PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.DOWNLOAD_PAPER_FORM_N1', externalURLs.n1FormUrl, true) }}
   {{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.AND') }} 
-  {{ externalLink(t, 'PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.FORM_N510', externalURLs.n510FormOutJurisdictionUrl, true) }}
+  {{ externalLink(t, 'PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.FORM_N510', externalURLs.n510Url, true) }}
   {{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.DEFENDANT_ADDRESS.COMPLETE_FORMS_AND_RETURN') }} 
 </p>
