+++ src/main/views/features/public/eligibility/not-eligible.njk
@@ -1,6 +1,6 @@
 {% extends "claim-details-tpl.njk" %}
-{% from "../../macro/page-title.njk" import setPageTitle %}
-{% from "../../macro/contact-us-for-help.njk" import contactUsForHelp %}
+{% from "../../../macro/page-title.njk" import setPageTitle %}
+{% from "../../../macro/contact-us-for-help.njk" import contactUsForHelp %}
 
 {% block pageTitle %}
   {{ setPageTitle('PAGES.NOT_ELIGIBLE_FOR_SERVICE.PAGE_TITLE') }}
@@ -11,11 +11,11 @@
     <div class="govuk-grid-column-two-thirds">
       <h1 class="govuk-heading-xl">{{ t('PAGES.NOT_ELIGIBLE_FOR_SERVICE.PAGE_TITLE') }}</h1>
       {% if reason === NotEligibleReason.CLAIM_VALUE_OVER_25000 %}
-        {% include "features/eligibility/not-eligible-includes/claim-value-over-25000.njk"%}
+        {% include "features/public/eligibility/not-eligible-includes/claim-value-over-25000.njk"%}
       {% elseif reason === NotEligibleReason.CLAIM_VALUE_NOT_KNOWN %}
-        {% include "features/eligibility/not-eligible-includes/claim-value-not-known.njk"%}
+        {% include "features/public/eligibility/not-eligible-includes/claim-value-not-known.njk"%}
       {% elseif reason === NotEligibleReason.DEFENDANT_ADDRESS %}
-        {% include "features/eligibility/not-eligible-includes/defendant-address.njk"%}
+        {% include "features/public/eligibility/not-eligible-includes/defendant-address.njk"%}
       {% endif %}
       {{ contactUsForHelp(t) }}
     </div>
