+++ src/test/unit/routes/features/public/eligibility/notEligibleController.test.ts
@@ -1,11 +1,11 @@
 import request from 'supertest';
-import {app} from '../../../../../main/app';
+import {app} from '../../../../../../main/app';
 import nock from 'nock';
 import config from 'config';
-import {NOT_ELIGIBLE_FOR_THIS_SERVICE_URL} from '../../../../../main/routes/urls';
-import {TestMessages} from '../../../../utils/errorMessageTestConstants';
+import {NOT_ELIGIBLE_FOR_THIS_SERVICE_URL} from '../../../../../../main/routes/urls';
+import {TestMessages} from '../../../../../utils/errorMessageTestConstants';
 
-jest.mock('../../../../../main/modules/oidc');
+jest.mock('../../../../../../main/modules/oidc');
 
 describe("You can't use this service", () => {
   const citizenRoleToken: string = config.get('citizenRoleToken');
