+++ src/test/utils/externalURLs.ts
@@ -5,5 +5,3 @@ export const citizensAdviceUrl = 'https://www.citizensadvice.org.uk/about-us/how
 export const governmentDepartmentsUrl = 'https://www.gov.uk/government/publications/serve-the-treasury-solicitor-with-legal-proceedings';
 export const tenancyServiceUrl = 'https://www.gov.uk/tenancy-deposit-protection/disputes-and-problems';
 export const feesHelpUri = 'https://www.gov.uk/get-help-with-court-fees';
-export const n1FormCprPart7Url = 'https://www.gov.uk/government/publications/form-n1-claim-form-cpr-part-7';
-export const n510FormOutJurisdictionUrl = 'https://www.gov.uk/government/publications/form-n510-service-out-of-the-jurisdiction';
